requirements.txt
customtkinter
openai-whisper